pyproject.toml
[project]
name = "arxivdigest-reimagined"
version = "0.2.0"
description = "Simplified arXiv digest generator with LLM support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "httpx[socks]>=0.28.1",
    "instructor>=1.12.0",
    "loguru>=0.7.3",
    "openai>=1.0,<2.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "pyyaml>=6.0.3",
]


[dependency-groups]
dev = ["ruff>=0.12.7"]


[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    # See: https://docs.astral.sh/ruff/rules/
    # Basic list from: https://docs.astral.sh/ruff/linter/#rule-selection
    "E",  # https://docs.astral.sh/ruff/rules/#error-e
    "F",  # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "UP", # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "B",  # https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
    "I",  # https://docs.astral.sh/ruff/rules/#isort-i
    # Other possibilities:
    # "D" # https://docs.astral.sh/ruff/rules/#pydocstyle-d
    # "Q" # https://docs.astral.sh/ruff/rules/#flake8-quotes-q
    # "COM" # https://docs.astral.sh/ruff/rules/#flake8-commas-com
    # "SIM", # https://docs.astral.sh/ruff/rules/#flake8-simplify-sim
]
ignore = [
    # Disable some rules that are overly pedantic. Add/remove as desired:
    "E501", # https://docs.astral.sh/ruff/rules/line-too-long/
    "E731", # https://docs.astral.sh/ruff/rules/lambda-assignment/
    # We use both ruff formatter and linter so some rules should always be disabled.
    # See: https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191",   # https://docs.astral.sh/ruff/rules/tab-indentation/
    "E111",   # https://docs.astral.sh/ruff/rules/indentation-with-invalid-multiple/
    "E114",   # https://docs.astral.sh/ruff/rules/indentation-with-invalid-multiple-comment/
    "E117",   # https://docs.astral.sh/ruff/rules/over-indented/
    "D206",   # https://docs.astral.sh/ruff/rules/docstring-tab-indentation/
    "D300",   # https://docs.astral.sh/ruff/rules/triple-single-quotes/
    "Q000",   # https://docs.astral.sh/ruff/rules/bad-quotes-inline-string/
    "Q001",   # https://docs.astral.sh/ruff/rules/bad-quotes-multiline-string/
    "Q002",   # https://docs.astral.sh/ruff/rules/bad-quotes-docstring/
    "Q003",   # https://docs.astral.sh/ruff/rules/avoidable-escaped-quote/
    "COM812", # https://docs.astral.sh/ruff/rules/missing-trailing-comma/
    "COM819", # https://docs.astral.sh/ruff/rules/prohibited-trailing-comma/
    "ISC002", # https://docs.astral.sh/ruff/rules/multi-line-implicit-string-concatenation/
]
